requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
pymongo==3.11.0
paho-mqtt==1.5.1
bcrypt==3.2.0
